Replace assert and __builtin_unreachable() with our own relatively cross-platform stuff

+++ src/AST/ASTPrinter.cpp
@@ -51,7 +51,7 @@ void ASTPrinter::visitStructDecl(StructDecl* decl, int indentationLevel, std::os
 }
 
 void ASTPrinter::visitScope(Scope* scope, int indentationLevel, std::ostream& stream) {
-    for(int i = 0; i < scope->nodes.size(); ++i) {
+    for(size_t i = 0; i < scope->nodes.size(); ++i) {
         visit(scope->nodes[i], indentationLevel + 1, stream);
         if(i != scope->nodes.size() - 1) stream << '\n';
     }
